+++ runtime/auth/aws-signing-crt/build.gradle.kts
@@ -8,14 +8,14 @@ extra["moduleName"] = "aws.smithy.kotlin.runtime.auth.awssigning.crt"
 
 kotlin {
     sourceSets {
-        getByName("jvmAndNativeMain") {
+        jvmMain {
             dependencies {
                 api(project(":runtime:auth:aws-signing-common"))
                 implementation(project(":runtime:crt-util"))
             }
         }
 
-        getByName("jvmAndNativeTest") {
+        jvmTest {
             dependencies {
                 implementation(project(":runtime:auth:aws-signing-tests"))
             }
