+++ runtime/auth/aws-signing-crt/jvm/test/aws/smithy/kotlin/runtime/auth/awssigning/crt/CrtAwsChunkedSourceTest.kt
@@ -7,6 +7,7 @@ package aws.smithy.kotlin.runtime.auth.awssigning.crt
 
 import aws.smithy.kotlin.runtime.auth.awssigning.AwsSigner
 import aws.smithy.kotlin.runtime.auth.awssigning.tests.AwsChunkedSourceTestBase
+import aws.smithy.kotlin.runtime.auth.awssigning.crt.CrtAwsSigner
 
 class CrtAwsChunkedSourceTest : AwsChunkedSourceTestBase() {
     override val signer: AwsSigner = CrtAwsSigner
