+++ runtime/auth/aws-signing-tests/native/src/aws/smithy/kotlin/runtime/auth/awssigning/tests/SigningSuiteTestBaseNative.kt
@@ -1,0 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package aws.smithy.kotlin.runtime.auth.awssigning.tests
+
+// FIXME Implement native tests
+actual abstract class SigningSuiteTestBase : HasSigner
