+++ runtime/auth/http-auth/common/test/aws/smithy/kotlin/runtime/http/auth/BearerTokenProviderChainTest.kt
@@ -9,8 +9,8 @@ import aws.smithy.kotlin.runtime.time.Instant
 import aws.smithy.kotlin.runtime.util.Attributes
 import aws.smithy.kotlin.runtime.util.emptyAttributes
 import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
+import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class BearerTokenProviderChainTest {
     private class TestTokenProvider(val token: String? = null) : BearerTokenProvider {
