+++ runtime/protocol/http-client/common/test/aws/smithy/kotlin/runtime/http/operation/SdkHttpOperationTest.kt
@@ -12,7 +12,7 @@ import aws.smithy.kotlin.runtime.telemetry.logging.loggingContext
 import aws.smithy.kotlin.runtime.telemetry.trace.traceSpan
 import io.kotest.matchers.string.shouldContain
 import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.Test
+import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
